[install.sh] --no-gitconfig option

+++ install.sh
@@ -16,14 +16,16 @@ mkdir -p ~/.config/karabiner
 mkdir -p ~/Library/Application\ Support/Cursor/User
 
 /bin/rm -f ~/.config/karabiner/karabiner.json
-/bin/rm -f ~/.gitconfig
 /bin/rm -f ~/.vimrc
 /bin/rm -f ~/.zshrc
 /bin/rm -f ~/Library/Application\ Support/Code/User/settings.json
 
-# create symlinks
+if [[ "$1" != "--no-gitconfig" ]]; then
+    /bin/rm -f ~/.gitconfig
+    ln -sf "${BASEDIR}/.gitconfig" ~/.gitconfig
+fi
+
 ln -sf "${BASEDIR}/karabiner.json" ~/.config/karabiner/karabiner.json
-# ln -sf "${BASEDIR}/.gitconfig" ~/.gitconfig
 ln -sf "${BASEDIR}/.vimrc" ~/.vimrc
 ln -sf "${BASEDIR}/.zshrc" ~/.zshrc
 ln -sf "${BASEDIR}/settings.json" ~/Library/Application\ Support/Code/User/settings.json
